feat: extract kotlinx.serialization to its own package
